Relocate tests for BindingIdentifier of strict fns
The modified test material was introduced via commit 73c001c on
2017-04-26. According to this project's `licenses.md` file, it
originated from the Acorn project. At that that date, the tests
containing the relevant productions were labeled as negative syntax
tests. However, the tests were introduced to this project as positive
syntax tests.

Although the new interpretation was technically correct in terms of the
ECMA262 specificatoin at that time, TC39 reached consensus in March 2018
to classify these productions as syntax errors.

Re-locate the relevant tests to align with the decision. Note that due
to the discrepency described above, no change to the source project is
necessary.