Remove 2 'early' tests re MV of HexDigits
Each of these tests is expected to raise a syntax error
because it has a string literal with a UnicodeEscapeSequence
that specifies an out-of-range code point.

Originally, this restriction was expressed as an early error:
    It is a Syntax Error if the MV of |HexDigits| > 0x10FFFF.
But PR tc39/ecma262#773 replaced it with an in-grammar restriction:
    CodePoint :: HexDigits [> but not if MV of HexDigits > 0x10FFFF ]

So you might expect this PR to move the tests from 'early' to 'fail',
but strangely, they already exist in 'fail'. In fact, these 2 tests
have existed in both 'early' and 'fail' since the start of the repo.
Until the merge of 773, only the 'early' copies were correct;
after, only the 'fail' copies are correct.
This PR removes the incorrect 'early' copies.

Resolves issue #15